requirements.txt
langchain==0.3.13
openai==1.58.1
chromadb==0.5.23
pypdf==5.1.0
streamlit==1.41.1
tiktoken==0.7.0
pytesseract==0.3.13
Pillow==11.1.0
python-docx==1.1.2
